requirements.txt
Flask
Werkzeug
azure-cognitiveservices-speech
azure-storage-blob
pytube
openai==0.28.0
openai-whisper
